+++ files/src/logic/backgroundestimation.cu
@@ -4,7 +4,7 @@
 
 //TODO: is lineptr_lab_candidate initialized with full 0??? (we need it)
 __global__ void check_background_kernel(ImageView<lab> in, ImageView<lab> currentBackground,
-                                            ImageView<lab> candidateBackground, ImageView<int> currentTimePixels,
+                                            ImageView<lab> candidateBackground, ImageView<uint8_t> currentTimePixels,
                                             ImageView<float> currentDistancePixels, int width, int height)
 {
 
@@ -64,7 +64,7 @@ __global__ void check_background_kernel(ImageView<lab> in, ImageView<lab> curren
 
 
 void check_background_cu(ImageView<lab> in, ImageView<lab> currentBackground,
-                            ImageView<lab> candidateBackground, ImageView<int> currentTimePixels,
+                            ImageView<lab> candidateBackground, ImageView<uint8_t> currentTimePixels,
                             ImageView<float> currentDistancePixels, int width, int height)
 {
     dim3 threadsPerBlock(32, 32);
